Refactor Clock_GetTimeMs into clock_posix.c
