+++ demos/mqtt/mqtt_demo_lightweight/CMakeLists.txt
@@ -6,13 +6,13 @@ target_sources(
     ${DEMO_NAME}
     PRIVATE
         "${DEMO_NAME}.c"
-        ${LOGGING_SOURCES}
 )
 
 target_link_libraries(
     ${DEMO_NAME}
     PRIVATE
         mqtt
+        clock_posix
         plaintext_posix
         reconnect_posix   
 )
