+++ demos/mqtt/mqtt_demo_plaintext/CMakeLists.txt
@@ -18,6 +18,7 @@ target_link_libraries(
     ${DEMO_NAME}
     PRIVATE
         mqtt
+        clock_posix
         plaintext_posix
         transport_reconnect_posix   
 )
