+++ platform/include/clock.h
@@ -1,5 +1,4 @@
 /*
- * IoT Platform V1.1.0
  * Copyright (C) 2020 Amazon.com, Inc. or its affiliates.  All Rights Reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -21,57 +20,23 @@
  */
 
 /**
- * @file iot_clock.h
+ * @file clock.h
  * @brief Time-related functions used by libraries in this SDK.
  */
 
-#ifndef IOT_CLOCK_H_
-#define IOT_CLOCK_H_
+#ifndef CLOCK_H_
+#define CLOCK_H_
 
 /* Standard includes. */
-#include <stdbool.h>
-#include <stddef.h>
 #include <stdint.h>
 
 /**
- * @functionspage{platform_clock,platform clock component,Clock}
- * - @functionname{platform_clock_function_gettimestring}
+ * @brief The timer query function.
+ *
+ * This function returns the elapsed time.
+ *
+ * @return Time in milliseconds.
  */
+uint32_t Clock_GetTimeMs( void );
 
-/**
- * @functionpage{IotClock_GetTimestring,platform_clock,gettimestring}
- */
-
-/**
- * @brief Generates a human-readable timestring, such as "01 Jan 2018 12:00".
- *
- * This function uses the system clock to generate a human-readable timestring.
- * This timestring is printed by the [logging functions](@ref logging_functions).
- *
- * @param[out] pBuffer A buffer to store the timestring in.
- * @param[in] bufferSize The size of `pBuffer`.
- * @param[out] pTimestringLength The actual length of the timestring stored in
- * `pBuffer`.
- *
- * @return `true` if a timestring was successfully generated; `false` otherwise.
- *
- * @warning The implementation of this function must not call any [logging functions]
- * (@ref logging_functions).
- *
- * <b>Example</b>
- * @code{c}
- * char timestring[ 32 ];
- * size_t timestringLength = 0;
- *
- * if( IotClock_GetTimestring( timestring, 32, &timestringLength ) == true )
- * {
- *     printf( "Timestring: %.*s", timestringLength, timestring );
- * }
- * @endcode
- */
-/* @[declare_platform_clock_gettimestring] */
-bool Clock_GetTimestring( char * pBuffer,
-                          size_t bufferSize,
-                          size_t * pTimestringLength );
-
-#endif /* ifndef IOT_CLOCK_H_ */
+#endif /* ifndef CLOCK_H_ */
