+++ platform/posix/clock_posix.c
@@ -20,15 +20,9 @@
  */
 
 /**
- * @file iot_clock_posix.c
- * @brief Implementation of the functions in iot_clock.h for POSIX systems.
+ * @file clock_posix.c
+ * @brief Implementation of the functions in clock.h for POSIX systems.
  */
-
-/* The config header is always included first. */
-#include "config.h"
-
-/* Standard includes. */
-#include <stdlib.h>
 
 /* POSIX include. Allow the default POSIX header to be overridden. */
 #ifdef POSIX_TIME_HEADER
@@ -40,64 +34,19 @@
 /* Platform clock include. */
 #include "clock.h"
 
-/* Configure logs for the functions in this file. */
-#ifdef LOG_LEVEL_PLATFORM
-    #define LIBRARY_LOG_LEVEL        LOG_LEVEL_PLATFORM
-#else
-    #ifdef LOG_LEVEL_GLOBAL
-        #define LIBRARY_LOG_LEVEL    LOG_LEVEL_GLOBAL
-    #else
-        #define LIBRARY_LOG_LEVEL    LOG_NONE
-    #endif
-#endif
-
-#define LIBRARY_LOG_NAME    ( "CLOCK" )
-#include "logging_setup.h"
-
 /*-----------------------------------------------------------*/
 
-/**
- * @brief The format of timestrings printed in logs.
- *
- * For more information on timestring formats, see [this link.]
- * (http://pubs.opengroup.org/onlinepubs/9699919799/functions/strftime.html)
- */
-#define TIMESTRING_FORMAT    ( "%F %R:%S" )
-
-/*-----------------------------------------------------------*/
-
-bool Clock_GetTimestring( char * pBuffer,
-                          size_t bufferSize,
-                          size_t * pTimestringLength )
+uint32_t Clock_GetTimeMs( void )
 {
-    bool status = true;
-    const time_t unixTime = time( NULL );
-    struct tm localTime = { 0 };
-    size_t timestringLength = 0;
+    uint32_t timeMs;
+    struct timespec timeSpec;
 
-    /* localtime_r is the thread-safe variant of localtime. Its return value
-     * should be the pointer to the localTime struct. */
-    if( localtime_r( &unixTime, &localTime ) != &localTime )
-    {
-        status = false;
-    }
+    /* Get the MONOTONIC time. */
+    clock_gettime( CLOCK_MONOTONIC, &timeSpec );
 
-    if( status == true )
-    {
-        /* Convert the localTime struct to a string. */
-        timestringLength = strftime( pBuffer, bufferSize, TIMESTRING_FORMAT, &localTime );
+    /* Calculate the milliseconds from timespec. */
+    timeMs = ( uint32_t ) ( timeSpec.tv_sec * 1000 )
+             + ( uint32_t ) ( timeSpec.tv_nsec / ( 1000 * 1000 ) );
 
-        /* Check for error from strftime. */
-        if( timestringLength == 0 )
-        {
-            status = false;
-        }
-        else
-        {
-            /* Set the output parameter. */
-            *pTimestringLength = timestringLength;
-        }
-    }
-
-    return status;
+    return timeMs;
 }
